[frontend] refractor&fix: MainNavBar, QueryBoard
Change name from NavBar to MainNavBar.
Move MainNavBar component to reside directly in App.js router.
Delete MainNavBar in Shop.js and Search.js to reduce redundancy.
Create the searchClick count state in App.js, allowing the QueryBoard
to listen to the number of time the Search button is clicked. This is
to prevent case when the user search a keyword, choose a filter but the
filter state remains when the user click the search button again.