[frontend] feat: Cart System
Enable the user to view item in their cart as well
as modify the quantity.
Enable the user to add an item into their cart
from shopping page.

ItemCard component is modified to include the
subcription to redux to get the userCode in order
to call the cart api from backend.