[backend] fix: Collision In Paying Cart
Fix an error which occur when the user tries to
checkout the cart but the item from the cart can
not move to the order item properly which create
a new order but still keep the cart content.

Error occurs because of Promise.all doesnt handle
the 3 trips to the database properly (SHOULDNT
MAKE MANY TRIP TO THE DATABASE IN THE FIRST PLACE)

Fix by creating a procedure in the database which
properly handle the action of inserting and
deleting.